chore: clean up + modularise
